requirements.txt
galois==0.4.6
gmpy2==2.2.1
matplotlib==3.10.3
networkx==3.1
numpy==2.3.1
Pillow==7.0.0
Pillow==11.3.0
pycryptodome==3.23.0
scipy==1.16.0
sympy==1.14.0